refactor(transaction-logger): replace Long with BigDecimal for money related fields

+++ services/transaction-logger/src/main/java/com/processing/transactionlogger/repository/TransactionRepository.java
@@ -6,13 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
 public interface TransactionRepository extends JpaRepository<Transaction, UUID>, JpaSpecificationExecutor<Transaction> {
     long countByStatus(TransactionStatus status);
     @Query("SELECT SUM(t.amount) FROM Transaction t")
-    long sumAmount();
+    BigDecimal sumAmount();
     long countByCreatedAtAfter(Instant since);
     @Query("SELECT AVG(t.processingTimeMs) FROM  Transaction t")
     double averageProcessingTimeMs();
